+++ zhihu/question.py
@@ -43,6 +43,7 @@ class Question(BaseZhihu):
         self._author = author
         self._creation_time = creation_time
         self._logs = None
+        self._deleted = None
 
     @property
     def url(self):
@@ -90,11 +91,6 @@ class Question(BaseZhihu):
         """
         return self.soup.prettify()
 
-    @html.deleter
-    def html(self):
-        if hasattr(self, '_html'):
-            del self._html
-
     @property
     @check_soup('_title')
     def title(self):
@@ -106,11 +102,6 @@ class Question(BaseZhihu):
         return self.soup.find('h2', class_='zm-item-title') \
             .text.replace('\n', '')
 
-    @title.deleter
-    def title(self):
-        if hasattr(self, '_title'):
-            del self._title
-
     @property
     @check_soup('_details')
     def details(self):
@@ -121,13 +112,8 @@ class Question(BaseZhihu):
         """
         return self.soup.find("div", id="zh-question-detail").div.text
 
-    @details.deleter
-    def details(self):
-        if hasattr(self, '_details'):
-            del self._details
-
     @property
-    @check_soup('_answers_num')
+    @check_soup('_answer_num')
     def answer_num(self):
         """获取问题答案数量.
 
@@ -145,11 +131,6 @@ class Question(BaseZhihu):
                 return 0
         return int(answer_num_block['data-num'])
 
-    @answer_num.deleter
-    def answer_num(self):
-        if hasattr(self, '_answer_num'):
-            del self._answer_num
-
     @property
     @check_soup('_follower_num')
     def follower_num(self):
@@ -164,11 +145,6 @@ class Question(BaseZhihu):
             return 0
         return int(follower_num_block.strong.text)
 
-    @follower_num.deleter
-    def follower_num(self):
-        if hasattr(self, '_follower_num'):
-            del self._follower_num
-
     @property
     @check_soup('_topics')
     def topics(self):
@@ -181,11 +157,6 @@ class Question(BaseZhihu):
         for topic in self.soup.find_all('a', class_='zm-item-tag'):
             topics_list.append(topic.text.replace('\n', ''))
         return topics_list
-
-    @topics.deleter
-    def topics(self):
-        if hasattr(self, '_topics'):
-            del self._topics
 
     @property
     def followers(self):
@@ -215,7 +186,11 @@ class Question(BaseZhihu):
         # TODO: 统一逻辑. 完全可以都用 _parse_answer_html 的逻辑替换
         if self._url.endswith('sort=created'):
             pager = self.soup.find('div', class_='zm-invite-pager')
-            max_page = int(pager.find_all('span')[-2].a.text)
+            if pager is None:
+                max_page = 1
+            else:
+                max_page = int(pager.find_all('span')[-2].a.text)
+
             for page in range(1, max_page + 1):
                 if page == 1:
                     soup = self.soup
@@ -366,11 +341,6 @@ class Question(BaseZhihu):
         time_string = soup.find_all('time')[0]['datetime']
         return datetime.strptime(time_string, "%Y-%m-%d %H:%M:%S")
 
-    @last_edit_time.deleter
-    def last_edit_time(self):
-        if hasattr(self, '_last_edit_time'):
-            del self._last_edit_time
-
     def _query_logs(self):
         if self._logs is None:
             gotten_feed_num = 20
@@ -399,14 +369,22 @@ class Question(BaseZhihu):
         :return: None
         """
         super().refresh()
-        del self.html
-        del self.title
-        del self.details
-        del self.answer_num
-        del self.follower_num
-        del self.topics
-        del self.last_edit_time
+        self._html = None
+        self._title = None
+        self._details = None
+        self._answer_num = None
+        self._follower_num = None
+        self._topics = None
+        self._last_edit_time = None
         self._logs = None
+
+    @property
+    @check_soup('_deleted')
+    def deleted(self):
+        """问题是否被删除, 被删除了返回 True, 未被删除返回 False
+        :return: True or False
+        """
+        return self._deleted
 
     def _parse_answer_html(self, answer_html, Author, Answer):
         soup = BeautifulSoup(answer_html)
@@ -430,4 +408,11 @@ class Question(BaseZhihu):
 
     def _get_content(self):
         # override base class's method cause we need self._url not self.url
-        return self._session.get(self._url).content
+        resp = self._session.get(self._url)
+
+        if resp.status_code == 404:
+            self._deleted = True
+        else:
+            self._deleted = False
+
+        return resp.content